pdf_parser/requirements_pdf_parser.txt
# Camelot (pdf parser - extract tables)
camelot-py==0.11.0
opencv-python==4.9.0.80
ghostscript==0.7